Add files via upload
This is a simple idea I used to move the mouse on the screen based on web camera input
Use a flashlight (White by default) to light up some pixels and the mouse shall move based on the approximate position of the light source
This is a simple implementation where the position of the light source is averaged by adding all the position vectors of very bright pixels and finding the average. surely it can be improved.

It requires the video library for processing 3
It uses java.awt.Robot class to control the mouse
A webcam is required to run the program
A dark environment gives a better approximation of the light source's position.